@@ -62,11 +62,14 @@
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.0.0",
-    "autoprefixer": "^9.1.3",
+    "@svgr/webpack": "2.4.1",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.0",
+    "babel-eslint": "^10.0.1",
     "babel-jest": "^23.4.2",
-    "babel-loader": "^8.0.0",
+    "babel-loader": "^8.0.4",
+    "babel-plugin-named-asset-import": "^0.2.2",
+    "babel-preset-react-app": "^5.0.3",
+    "bfj": "6.1.1",
     "bignumber.js": "^7.2.1",
     "browserslist": "^4.1.1",
     "canvas": "^1.6.11",
@@ -75,7 +78,7 @@
     "css-loader": "^1.0.0",
     "dotenv": "^6.0.0",
     "dotenv-expand": "^4.2.0",
-    "eslint": "^5.4.0",
+    "eslint": "^5.6.1",
     "eslint-config-oceanprotocol": "^1.1.1",
     "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-react": "^7.0.2",
@@ -87,7 +90,7 @@
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-standard": "^4.0.0",
     "file-loader": "^2.0.0",
-    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^4.0.0-beta.1",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^23.5.0",
     "mini-css-extract-plugin": "^0.4.3",
@@ -95,20 +98,21 @@
     "npm-run-all": "^4.1.3",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "postcss-loader": "^3.0.0",
-    "react-dev-utils": "6.0.0-next.a671462c",
+    "postcss-preset-env": "^6.1.0",
+    "postcss-safe-parser": "^4.0.1",
+    "react-dev-utils": "^6.0.4",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.0",
     "stylelint": "^9.5.0",
     "stylelint-config-bigchaindb": "^1.2.1",
     "stylelint-config-css-modules": "^1.3.0",
     "stylelint-config-standard": "^18.2.0",
-    "sw-precache-webpack-plugin": "^0.11.5",
     "terser-webpack-plugin": "^1.1.0",
-    "thread-loader": "^1.2.0",
     "url-loader": "^1.1.1",
     "webpack": "^4.19.1",
     "webpack-dev-server": "^3.1.8",
-    "webpack-manifest-plugin": "^2.0.3"
+    "webpack-manifest-plugin": "^2.0.3",
+    "workbox-webpack-plugin": "^3.6.2"
   },
   "browserslist": {
     "development": [
@@ -118,7 +122,7 @@
       ">1%",
       "last 4 versions",
       "Firefox ESR",
-      "not ie < 11"
+      "not ie <= 11"
     ]
   },
   "jest": {
