@@ -17,14 +17,14 @@ const isLocalhost = Boolean(window.location.hostname === 'localhost' ||
     // 127.0.0.1/8 is considered localhost for IPv4.
     window.location.hostname.match(/^127(?:\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}$/)) // eslint-disable-line security/detect-unsafe-regex
 
-export default function register() {
+export function register(config) {
     if (process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
-    // The URL constructor is available in all browsers that support SW.
+        // The URL constructor is available in all browsers that support SW.
         const publicUrl = new URL(process.env.PUBLIC_URL, window.location)
         if (publicUrl.origin !== window.location.origin) {
             // Our service worker won't work if PUBLIC_URL is on a different origin
             // from what our page is served on. This might happen if a CDN is used to
-            // serve assets; see https://github.com/facebookincubator/create-react-app/issues/2374
+            // serve assets; see https://github.com/facebook/create-react-app/issues/2374
             return
         }
 
@@ -32,24 +32,26 @@ export default function register() {
             const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`
 
             if (isLocalhost) {
-                // This is running on localhost. Lets check if a service worker still exists or not.
-                checkValidServiceWorker(swUrl)
+                // This is running on localhost. Let's check if a service worker still exists or not.
+                checkValidServiceWorker(swUrl, config)
 
                 // Add some additional logging to localhost, pointing developers to the
                 // service worker/PWA documentation.
                 navigator.serviceWorker.ready.then(() => {
-                    console.log('This web app is being served cache-first by a service ' + // eslint-disable-line no-console
-              'worker. To learn more, visit https://goo.gl/SC7cgQ')
+                    console.log(
+                        'This web app is being served cache-first by a service ' +
+                    'worker. To learn more, visit https://goo.gl/SC7cgQ'
+                    )
                 })
             } else {
                 // Is not local host. Just register service worker
-                registerValidSW(swUrl)
+                registerValidSW(swUrl, config)
             }
         })
     }
 }
 
-function registerValidSW(swUrl) {
+function registerValidSW(swUrl, config) {
     navigator.serviceWorker
         .register(swUrl)
         .then(registration => {
@@ -62,12 +64,22 @@ function registerValidSW(swUrl) {
                             // the fresh content will have been added to the cache.
                             // It's the perfect time to display a "New content is
                             // available; please refresh." message in your web app.
-                            console.log('New content is available; please refresh.') // eslint-disable-line no-console
+                            console.log('New content is available; please refresh.')
+
+                            // Execute callback
+                            if (config && config.onUpdate) {
+                                config.onUpdate(registration)
+                            }
                         } else {
                             // At this point, everything has been precached.
                             // It's the perfect time to display a
                             // "Content is cached for offline use." message.
-                            console.log('Content is cached for offline use.') // eslint-disable-line no-console
+                            console.log('Content is cached for offline use.')
+
+                            // Execute callback
+                            if (config && config.onSuccess) {
+                                config.onSuccess(registration)
+                            }
                         }
                     }
                 }
@@ -78,14 +90,14 @@ function registerValidSW(swUrl) {
         })
 }
 
-function checkValidServiceWorker(swUrl) {
+function checkValidServiceWorker(swUrl, config) {
     // Check if the service worker can be found. If it can't reload the page.
     fetch(swUrl)
         .then(response => {
             // Ensure service worker exists, and that we really are getting a JS file.
             if (
                 response.status === 404 ||
-        response.headers.get('content-type').indexOf('javascript') === -1
+              response.headers.get('content-type').indexOf('javascript') === -1
             ) {
                 // No service worker found. Probably a different app. Reload the page.
                 navigator.serviceWorker.ready.then(registration => {
@@ -95,11 +107,13 @@ function checkValidServiceWorker(swUrl) {
                 })
             } else {
                 // Service worker found. Proceed as normal.
-                registerValidSW(swUrl)
+                registerValidSW(swUrl, config)
             }
         })
         .catch(() => {
-            console.log('No internet connection found. App is running in offline mode.') // eslint-disable-line no-console
+            console.log(
+                'No internet connection found. App is running in offline mode.'
+            )
         })
 }
 
